requirements.txt
Flask
Flask-SocketIO
python-socketio[client]
PyJWT
eventlet
redis
